Relocate macros to another mod
