+++ src/core/mod.rs
@@ -1,1 +1,1 @@
-mod syscall_handler;
+pub mod syscall_handler;
