+++ src/lib.rs
@@ -1,1 +1,2 @@
-mod core;
+pub mod core;
+pub mod utils;
