+++ src/utils.rs
@@ -1,0 +1,81 @@
+#[cfg(test)]
+#[macro_use]
+pub mod test_utils {
+    #[macro_export]
+    macro_rules! add_segments {
+        ($vm:expr, $n:expr) => {
+            for _ in 0..$n {
+                $vm.add_memory_segment();
+            }
+        };
+    }
+    pub(crate) use add_segments;
+
+    #[macro_export]
+    macro_rules! exec_scopes_ref {
+        () => {
+            &mut ExecutionScopes::new()
+        };
+    }
+    pub(crate) use exec_scopes_ref;
+
+    #[macro_export]
+    macro_rules! run_hint {
+        ($vm:expr, $ids_data:expr, $hint_code:expr, $exec_scopes:expr, $constants:expr) => {{
+            let hint_data = HintProcessorData::new_default($hint_code.to_string(), $ids_data);
+            let hint_processor = BuiltinHintProcessor::new_empty();
+            hint_processor.execute_hint(&mut $vm, $exec_scopes, &any_box!(hint_data), $constants)
+        }};
+        ($vm:expr, $ids_data:expr, $hint_code:expr, $exec_scopes:expr) => {{
+            let hint_data = HintProcessorData::new_default($hint_code.to_string(), $ids_data);
+            let hint_processor = BuiltinHintProcessor::new_empty();
+            hint_processor.execute_hint(
+                &mut $vm,
+                $exec_scopes,
+                &any_box!(hint_data),
+                &HashMap::new(),
+            )
+        }};
+        ($vm:expr, $ids_data:expr, $hint_code:expr) => {{
+            let hint_data = HintProcessorData::new_default($hint_code.to_string(), $ids_data);
+            let hint_processor = BuiltinHintProcessor::new_empty();
+            hint_processor.execute_hint(
+                &mut $vm,
+                exec_scopes_ref!(),
+                &any_box!(hint_data),
+                &HashMap::new(),
+            )
+        }};
+    }
+    pub(crate) use run_hint;
+
+    #[macro_export]
+    macro_rules! run_syscall_hint {
+        ($vm:expr, $ids_data:expr, $hint_code:expr, $exec_scopes:expr, $constants:expr) => {{
+            let hint_data = HintProcessorData::new_default($hint_code.to_string(), $ids_data);
+            let hint_processor = BuiltinHintProcessor::new_empty();
+            hint_processor.execute_hint(&mut $vm, $exec_scopes, &any_box!(hint_data), $constants)
+        }};
+        ($vm:expr, $ids_data:expr, $hint_code:expr, $exec_scopes:expr) => {{
+            let hint_data = HintProcessorData::new_default($hint_code.to_string(), $ids_data);
+            let hint_processor = BuiltinHintProcessor::new_empty();
+            hint_processor.execute_hint(
+                &mut $vm,
+                $exec_scopes,
+                &any_box!(hint_data),
+                &HashMap::new(),
+            )
+        }};
+        ($vm:expr, $ids_data:expr, $hint_code:expr) => {{
+            let hint_data = HintProcessorData::new_default($hint_code.to_string(), $ids_data);
+            let hint_processor = SyscallHintProcessor::new_empty();
+            hint_processor.execute_hint(
+                &mut $vm,
+                exec_scopes_ref!(),
+                &any_box!(hint_data),
+                &HashMap::new(),
+            )
+        }};
+    }
+    pub(crate) use run_syscall_hint;
+}
